+++ imml/visualize/__init__.py
@@ -3,3 +3,4 @@
 from .plot_pid import plot_pid
 from .plot_missing_modality import plot_missing_modality
 from .plot_summary import plot_summary
+from .plot_combinations import plot_combinations