+++ imml/visualize/plot_missing_modality.py
@@ -6,6 +6,7 @@ import pandas as pd
 from matplotlib import pyplot as plt
 
 from ..impute import get_observed_mod_indicator
+from ..utils import check_Xs_y
 
 
 def plot_missing_modality(Xs, ax: matplotlib.axes.Axes = None, figsize: tuple = None, sort: bool = True):
@@ -50,12 +51,7 @@ def plot_missing_modality(Xs, ax: matplotlib.axes.Axes = None, figsize: tuple = 
     >>> Xs = transformer.fit_transform(Xs)
     >>> plot_missing_modality(Xs=Xs)
     """
-    if not isinstance(Xs, list):
-        raise ValueError(f"Invalid Xs. It must be a list. A {type(Xs)} was passed.")
-    if any(len(X) == 0 for X in Xs):
-        raise ValueError("Invalid Xs. All elements must have at least one sample.")
-    if len(set(len(X) for X in Xs)) > 1:
-        raise ValueError("Invalid Xs. All elements must have the same number of samples.")
+    Xs = check_Xs_y(Xs=Xs)
     if (ax is not None) and (not isinstance(ax, matplotlib.axes.Axes)):
         raise ValueError(f"Invalid ax. It must be a matplotlib.axes.Axes. A {type(ax)} was passed.")
     if (figsize is not None) and (not isinstance(figsize, tuple)):